Ship Active Mission Location updated (not tested)

+++ src/main/java/us/dev/shipandcargo/Vo/ShipIdelLocationVo.java
@@ -16,7 +16,7 @@ public class ShipIdelLocationVo {
 
     private Long imo;
     private String portId;
-    private int uploaderId;
+    private Long uploaderId;
 
     @JsonFormat(shape = JsonFormat.Shape.NUMBER)
     private Date createdAt;
